# Api/wwwroot/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Login API",
    "version": "1.0.0"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Verify login credentials",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserModel" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserModel" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "409": { "description": "User exists" }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "summary": "Get user by username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserModel" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete user",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "summary": "Update user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserModel" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/users/{id}/description": {
      "put": {
        "summary": "Update description",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Updated" },
          "404": { "description": "Not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserModel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "password": { "type": "string" },
          "description": { "type": "string" }
        }
      }
    }
  }
}